First version of Hub to translate Harmony commands into MQTT messages
The Hub primarily intends to provide a bridge between the Harmony remote and Home Assistant by translating the nRF24L01+ radio packets sent by the Harmony remote to easily interpreted MQTT messages. The name of individual mqtt messages (payloads) can be set in the command list, and Harmony remote buttons can be configured to be one of three functional types to provide a balance between responsiveness and expressivity.